backend/requirements.txt
flask==2.2.3
google-generativeai==0.2.0  # or whichever version you use
transformers==4.28.1
torch==2.6.0  # 2.0.1
pytest==7.4.0
requests==2.31.0
